Changed skill db to use SkillRank

+++ src/ChannelServer/Network/Handlers/Skills.cs
@@ -36,7 +36,7 @@ namespace Aura.Channel.Network.Handlers
 			var skill = creature.Skills.GetSafe(skillId);
 			if (!skill.IsRankable) goto L_Fail;
 
-			var nextRank = skill.Data.GetRankData((int)skill.Info.Rank + 1, creature.RaceId);
+			var nextRank = skill.Data.GetRankData(skill.Info.Rank + 1, creature.RaceId);
 			if (nextRank == null)
 			{
 				Log.Warning("Player '{0:X16}' tried to advance skill '{1}' to unknown rank '{2}'.", creature.EntityId, skill.Info.Id, skill.Info.Rank + 1);
